+++ lib/model/arrayList.js
@@ -42,9 +42,18 @@ var ArrayList = createClass({
       return this._backing.length;
     },
 
+    /**
+     * Validate values for this array. Intended to be overwritten.
+     * @method  _validateValue
+     * @param   {Any} val
+     * @protected
+     * @return  {Boolea} valid?
+     */
+    /* eslint-disable */
     _validateValue: function(val) {
       return true;
     }
+    /* eslint-enable */
   }
 });
