+++ lib/model/reroutePointsList.js
